// src/types.ts
export interface StudentData {
  year: number;
  courses_enrolled: number;
  hist_MATH101_coursework: number;
  hist_MATH101_exam: number;
  hist_MATH101_final: number;
  hist_MATH101_grade_point: number;
  hist_CS101_coursework: number;
  hist_CS101_exam: number;
  hist_CS101_final: number;
  hist_CS101_grade_point: number;
  hist_STAT101_coursework: number;
  hist_STAT101_exam: number;
  hist_STAT101_final: number;
  hist_STAT101_grade_point: number;
  hist_ECON101_coursework: number;
  hist_ECON101_exam: number;
  hist_ECON101_final: number;
  hist_ECON101_grade_point: number;
  hist_MATH201_coursework: number;
  hist_MATH201_exam: number;
  hist_MATH201_final: number;
  hist_MATH201_grade_point: number;
  hist_CS120_coursework: number;
  hist_CS120_exam: number;
  hist_CS120_final: number;
  hist_CS120_grade_point: number;
  MATH301_coursework_avg: number;
  MATH301_final_exam: number;
  MATH301_exam_coursework_diff: number;
  MATH301_dept_improvement: number;
  MATH301_dept_gpa_improvement: number;
  MATH301_historical_improvement: number;
  MATH301_gpa_improvement: number;
  STAT240_coursework_avg: number;
  STAT240_final_exam: number;
  STAT240_exam_coursework_diff: number;
  STAT240_dept_improvement: number;
  STAT240_dept_gpa_improvement: number;
  STAT240_historical_improvement: number;
  STAT240_gpa_improvement: number;
  CS220_coursework_avg: number;
  CS220_final_exam: number;
  CS220_exam_coursework_diff: number;
  CS220_dept_improvement: number;
  CS220_dept_gpa_improvement: number;
  CS220_historical_improvement: number;
  CS220_gpa_improvement: number;
  ECON210_coursework_avg: number;
  ECON210_final_exam: number;
  ECON210_exam_coursework_diff: number;
  ECON210_dept_improvement: number;
  ECON210_dept_gpa_improvement: number;
  ECON210_historical_improvement: number;
  ECON210_gpa_improvement: number;
  DS350_coursework_avg: number;
  DS350_final_exam: number;
  DS350_exam_coursework_diff: number;
  DS350_historical_improvement: number;
  DS350_gpa_improvement: number;
  CS340_coursework_avg: number;
  CS340_final_exam: number;
  CS340_exam_coursework_diff: number;
  CS340_dept_improvement: number;
  CS340_dept_gpa_improvement: number;
  CS340_historical_improvement: number;
  CS340_gpa_improvement: number;
  MATH330_coursework_avg: number;
  MATH330_final_exam: number;
  MATH330_exam_coursework_diff: number;
  MATH330_dept_improvement: number;
  MATH330_dept_gpa_improvement: number;
  MATH330_historical_improvement: number;
  MATH330_gpa_improvement: number;
  STAT330_coursework_avg: number;
  STAT330_final_exam: number;
  STAT330_exam_coursework_diff: number;
  STAT330_dept_improvement: number;
  STAT330_dept_gpa_improvement: number;
  STAT330_historical_improvement: number;
  STAT330_gpa_improvement: number;
  ECON315_coursework_avg: number;
  ECON315_final_exam: number;
  ECON315_exam_coursework_diff: number;
  ECON315_dept_improvement: number;
  ECON315_dept_gpa_improvement: number;
  ECON315_historical_improvement: number;
  ECON315_gpa_improvement: number;
  FIN300_coursework_avg: number;
  FIN300_final_exam: number;
  FIN300_exam_coursework_diff: number;
  FIN300_historical_improvement: number;
  FIN300_gpa_improvement: number;
  semester_gpa: number;
  suspicious_improvements_count: number;
  semester_gpa_improvement: number;
}

export interface PredictionResult {
  anomaly_probability: number;
  flagged_as_cheating: boolean;
  risk_level: string;
  anomaly_score: number;
}

export interface StudentInfo {
  student_id: string;
  gender: string;
  program: string;
  year: number;
  semester: string;
}